+++ digiroad2-oracle/src/main/scala/fi/liikennevirasto/digiroad2/asset/oracle/LRMPosition.scala
@@ -1,6 +1,5 @@
 package fi.liikennevirasto.digiroad2.asset.oracle
 
-import oracle.spatial.geometry.JGeometry
 import fi.liikennevirasto.digiroad2.Point
 
-case class LRMPosition(id: Long, startMeasure: Int, endMeasure: Int, point: Option[Point])
+case class LRMPosition(id: Long, startMeasure: Double, endMeasure: Double, point: Option[Point])
